+++ src/components/ProductUpload/ImageUpload.jsx
@@ -4,8 +4,8 @@ import { Card } from "@/components/ui/card";
 
 const ImageUpload = ({ index, image, onChange }) => {
   return (
-    <div className="relative">
-      <Card className="border-dashed border-2 border-gray-400 h-32 flex items-center justify-center cursor-pointer">
+    <div className="relative aspect-square">
+      <Card className="w-full h-full flex items-center border-2 border-gray-400  justify-center cursor-pointer">
         <input
           type="file"
           accept="image/*"
@@ -17,10 +17,10 @@ const ImageUpload = ({ index, image, onChange }) => {
           <img
             src={image}
             alt={`업로드된 이미지 ${index + 1}`}
-            className="h-full object-cover rounded"
+            className="w-full h-full object-cover rounded"
           />
         ) : (
-          <span className="text-gray-500">이미지 업로드</span>
+          <span className="text-gray-400">이미지 업로드</span>
         )}
       </Card>
     </div>
